Add files via upload
This Python project is built using pygame library. Pygame is a great library that will allow us to create the window and draw images and shapes on the window. This way we will capture mouse coordinates and identify the block where we need to mark ‘X’ or ‘O’. Then we will check if the user wins the game or not.
Steps to build Tic Tac Toe program in Python:
1.Create the display window for our game.
2.Draw the grid on the canvas where we will play Tic Tac Toe.
3.Draw the status bar below the canvas to show which player’s turn is it and who wins the game.
4.When someone wins the game or the game is a draw then we reset the game.